Review fixes: memory, precision, validation, and cleanups
- Store tokenizations as int64 numpy arrays and batch-tokenize in one
  call; Python int lists cost ~28 B/token, which at corpus scale (250k x
  ~1k tokens) is multiple GB of avoidable RAM.
- Accumulate perplexity/entropy/outlier means in float64; float32
  accumulation over long texts was a precision regression vs the old
  Python-float sums.
- Restore lost input validation: topp in (0,1], topk >= 1 (previously
  ValueError-checked in the metric functions), plus sizing checks and a
  topk<=vocab guard; fail fast with a clear error when a configured
  columns_to_score entry is missing from the dataset.
- Reuse the top-p full sort for the top-k threshold (drop redundant
  torch.topk); hoist the chunk-concat helper to module level; remove a
  dead del loop in the teardown path and document the shared-list free;
  table-driven metric dispatch in llm_stats instead of an if/elif chain.

+++ src/fastdetector/statistics/statistics_llm.py
@@ -33,7 +33,7 @@ def perplexity(token_lps: np.ndarray) -> float:
     """
     if token_lps.size == 0:
         return float("nan")
-    avg_logprob = float(np.mean(token_lps))
+    avg_logprob = float(np.mean(token_lps, dtype=np.float64))
     try:
         return math.exp(-avg_logprob)
     except OverflowError:
@@ -51,7 +51,7 @@ def mean_entropy(entropies: np.ndarray) -> float:
     """
     if entropies.size == 0:
         return 0.0
-    return float(np.mean(entropies))
+    return float(np.mean(entropies, dtype=np.float64))
 
 
 def outlier_percentage(outlier_flags: np.ndarray) -> float:
@@ -65,7 +65,7 @@ def outlier_percentage(outlier_flags: np.ndarray) -> float:
     """
     if outlier_flags.size == 0:
         return float("nan")
-    return float(np.mean(outlier_flags))
+    return float(np.mean(outlier_flags, dtype=np.float64))
 
 
 def fastdetectgpt_score(
